pyproject.toml
[project]
name = "langchain-streamlit-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "langchain>=0.3.25",
    "langchain-chroma>=0.2.4",
    "langchain-community>=0.3.25",
    "langchain-core>=0.3.65",
    "langchain-openai>=0.3.24",
    "langchain-text-splitters>=0.3.8",
    "protobuf==3.20.*",
    "python-dotenv>=1.1.0",
    "streamlit>=1.46.0",
]
